+++ src/modules/all.ts
@@ -42,3 +42,6 @@ export * from './esl-animate/core';
 
 // Related Target Mixin
 export * from './esl-related-target/core';
+
+// Share
+export * from './esl-share/core';