+++ src/modules/esl-event-listener/core.ts
@@ -1,8 +1,9 @@
 export * from './core/api';
+export * from './core/targets/resize.adapter';
+export * from './core/targets/decorated.target';
 export type {
   ESLListenerHandler,
   ESLListenerCriteria,
-  ESLListenerEventMap,
   ESLListenerDescriptor,
   ESLListenerDescriptorFn,
   ESLListenerDescriptorExt,