+++ src/modules/esl-media-query/core/esl-media-rule-list.ts
@@ -151,7 +151,7 @@ export class ESLMediaRuleList<T = any> extends SyntheticEventTarget {
   public override addEventListener(type: 'change', callback: EventListener): void;
   public override addEventListener(type: any, callback: EventListener = type): void {
     super.addEventListener(type, callback);
-    if (this.hasEventListener(1)) return;
+    if (this.getEventListeners('change').length > 1) return;
     this._value = this.computedValue;
     this.rules.forEach((rule) => rule.addEventListener(this._onMatchChanged));
   }