+++ src/modules/esl-toggleable/core/esl-toggleable.ts
@@ -7,7 +7,7 @@ import {sequentialUID} from '../../esl-utils/misc/uid';
 import {DeviceDetector} from '../../esl-utils/environment/device-detector';
 import {DelayedTask} from '../../esl-utils/async/delayed-task';
 import {ESLBaseElement} from '../../esl-base-element/core';
-import {isMatches} from '../../esl-utils/dom/traversing';
+import {findParent, isMatches} from '../../esl-utils/dom/traversing';
 
 /** Default Toggleable action params type definition */
 export interface ESLToggleableActionParams {
@@ -39,7 +39,7 @@ export interface ESLToggleableRequestDetails extends ESLToggleableActionParams {
   match?: string | ((target: Element) => boolean);
 }
 
-/** @deprecated alias for ESLToggleableActionParams */
+/** @deprecated alias for ESLToggleableActionParams, will be removed in 5.0.0*/
 export type ToggleableActionParams = ESLToggleableActionParams;
 
 const activators: WeakMap<ESLToggleable, HTMLElement | undefined> = new WeakMap();
@@ -78,10 +78,23 @@ export class ESLToggleable extends ESLBaseElement {
   /** Event to dispatch when toggleable group has changed */
   @prop('esl:change:group') public GROUP_CHANGED_EVENT: string;
 
-  /** CSS class to add on the body element */
+  /**
+   * CSS class (supports {@link CSSClassUtils}) to add on the body element
+   * */
   @attr() public bodyClass: string;
-  /** CSS class to add when the Toggleable is active */
+  /** CSS class (supports {@link CSSClassUtils}) to add when the Toggleable is active */
   @attr({defaultValue: 'open'}) public activeClass: string;
+
+  /**
+   * CSS class (supports {@link CSSClassUtils}) to add/remove on the container
+   * defined by {@link containerActiveClassTarget}
+   */
+  @attr() public containerActiveClass: string;
+  /**
+   * Selector for the closest parent element to add/remove {@link containerActiveClass}
+   * (default: `*` direct parent)
+   */
+  @attr({defaultValue: '*'}) public containerActiveClassTarget: string;
 
   /** Toggleable group meta information to organize groups */
   @attr({name: 'group'}) public groupName: string;
@@ -121,8 +134,7 @@ export class ESLToggleable extends ESLBaseElement {
   protected override connectedCallback(): void {
     super.connectedCallback();
     if (!this.id && !this.noAutoId) {
-      const tag = (this.constructor as typeof ESLToggleable).is;
-      this.id = sequentialUID(tag, tag + '-');
+      this.id = sequentialUID(this.baseTagName, this.baseTagName + '-');
     }
     this.initiallyOpened = this.hasAttribute('open');
     this.setInitialState();
@@ -133,7 +145,7 @@ export class ESLToggleable extends ESLBaseElement {
     activators.delete(this);
   }
 
-  protected attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
+  protected override attributeChangedCallback(attrName: string, oldVal: string, newVal: string): void {
     if (!this.connected || newVal === oldVal) return;
     switch (attrName) {
       case 'open':
@@ -141,7 +153,7 @@ export class ESLToggleable extends ESLBaseElement {
         this.toggle(this.open, {initiator: 'attribute', showDelay: 0, hideDelay: 0});
         break;
       case 'group':
-        this.$$fire(this.GROUP_CHANGED_EVENT,  {
+        this.$$fire(this.GROUP_CHANGED_EVENT, {
           detail: {oldGroupName: oldVal, newGroupName: newVal}
         });
         break;
@@ -224,6 +236,11 @@ export class ESLToggleable extends ESLBaseElement {
   protected onShow(params: ESLToggleableActionParams): void {
     CSSClassUtils.add(this, this.activeClass);
     CSSClassUtils.add(document.body, this.bodyClass, this);
+    if (this.containerActiveClass) {
+      const $container = findParent(this, this.containerActiveClassTarget);
+      $container && CSSClassUtils.add($container, this.containerActiveClass, this);
+    }
+
     this.updateA11y();
     this.$$fire(this.REFRESH_EVENT); // To notify other components about content change
   }
@@ -236,6 +253,10 @@ export class ESLToggleable extends ESLBaseElement {
   protected onHide(params: ESLToggleableActionParams): void {
     CSSClassUtils.remove(this, this.activeClass);
     CSSClassUtils.remove(document.body, this.bodyClass, this);
+    if (this.containerActiveClass) {
+      const $container = findParent(this, this.containerActiveClassTarget);
+      $container && CSSClassUtils.remove($container, this.containerActiveClass, this);
+    }
     this.updateA11y();
   }
 
@@ -310,13 +331,25 @@ export class ESLToggleable extends ESLBaseElement {
   @listen({auto: false, event: 'mouseenter'})
   protected _onMouseEnter(e: MouseEvent): void {
     const hideDelay = this._trackHoverDelay;
-    const baseParams: ESLToggleableActionParams = {initiator: 'mouseenter', trackHover: true, activator: this.activator, event: e, hideDelay};
+    const baseParams: ESLToggleableActionParams = {
+      initiator: 'mouseenter',
+      trackHover: true,
+      activator: this.activator,
+      event: e,
+      hideDelay
+    };
     this.show(Object.assign(baseParams, this.trackHoverParams));
   }
   @listen({auto: false, event: 'mouseleave'})
   protected _onMouseLeave(e: MouseEvent): void {
     const hideDelay = this._trackHoverDelay;
-    const baseParams: ESLToggleableActionParams = {initiator: 'mouseleave', trackHover: true, activator: this.activator, event: e, hideDelay};
+    const baseParams: ESLToggleableActionParams = {
+      initiator: 'mouseleave',
+      trackHover: true,
+      activator: this.activator,
+      event: e,
+      hideDelay
+    };
     this.hide(Object.assign(baseParams, this.trackHoverParams));
   }
 
@@ -345,6 +378,7 @@ declare global {
   export interface ESLLibrary {
     Toggleable: typeof ESLToggleable;
   }
+
   export interface HTMLElementTagNameMap {
     'esl-toggleable': ESLToggleable;
   }